Log errors from telemetry events
