+++ server/command.go
@@ -675,7 +675,7 @@ If you ran |v2revert| unintentionally and would like to continue using the curre
 	message = preMessage + message
 	message = strings.ReplaceAll(message, "|", "`")
 
-	p.track("v2RevertSubmitted", header.UserId)
+	p.TrackUserEvent("v2RevertSubmitted", header.UserId, nil)
 
 	return p.responsef(header, message)
 }
